BUGFIX: redirect on deleting current page
Move intialization of all serverFeedbackHandlers into one place, so they
don't override each other by accident.
